backend/requirements.txt
absl-py==1.4.0
aiosqlite==0.19.0
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
cmake==3.26.3
cryptography==40.0.2
databases==0.6.2
dm-tree==0.1.8
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
etils==1.2.0
fastapi==0.95.1
fastapi-pagination==0.12.2
filelock==3.12.0
flax==0.6.9
fsspec==2023.4.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.14.1
hyperframe==6.0.1
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
jaxlib==0.4.7
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.3
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.2.1
orjson==3.8.11
ormar==0.12.1
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
protobuf==4.22.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
qdrant-client==1.1.6
redis==4.5.4
regex==2023.5.5
requests==2.30.0
rich==13.3.5
rq==1.14.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.26.1
sympy==1.11.1
tensorstore==0.1.36
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
whisper-jax @ git+https://github.com/sanchit-gandhi/whisper-jax.git@0d3bc54c8c673d71732ba6bc20fa0323b1101fa6
xmltodict==0.13.0
zipp==3.15.0
